+++ frontend/src/client/types.gen.ts
@@ -2474,6 +2474,7 @@ export type FeatureFlag =
   | "case-durations"
   | "case-tasks"
   | "case-triggers"
+  | "rbac"
 
 /**
  * Response model for feature flags.
@@ -9270,6 +9271,8 @@ export type VcsDeleteGithubAppCredentialsResponse = void
 export type VcsGetGithubAppCredentialsStatusResponse =
   GitHubAppCredentialsStatus
 
+export type UsersGetMyScopesResponse = UserScopesRead
+
 export type RbacListScopesData = {
   /**
    * Include system/registry scopes
@@ -9446,8 +9449,6 @@ export type RbacDeleteUserAssignmentData = {
 }
 
 export type RbacDeleteUserAssignmentResponse = void
-
-export type UsersGetMyScopesResponse = UserScopesRead
 
 export type UsersUsersCurrentUserResponse = UserRead
 
@@ -13788,6 +13789,16 @@ export type $OpenApiTs = {
       }
     }
   }
+  "/users/me/scopes": {
+    get: {
+      res: {
+        /**
+         * Successful Response
+         */
+        200: UserScopesRead
+      }
+    }
+  }
   "/rbac/scopes": {
     get: {
       req: RbacListScopesData
@@ -14139,16 +14150,6 @@ export type $OpenApiTs = {
          * Validation Error
          */
         422: HTTPValidationError
-      }
-    }
-  }
-  "/users/me/scopes": {
-    get: {
-      res: {
-        /**
-         * Successful Response
-         */
-        200: UserScopesRead
       }
     }
   }
