+++ packages/tracecat-ee/tracecat_ee/rbac/__init__.py
@@ -1,0 +1,1 @@
+"""Enterprise Edition RBAC management module."""
