+++ packages/tracecat-ee/tracecat_ee/rbac/service.py
@@ -14,6 +14,7 @@ from tracecat.db.models import (
     Group,
     GroupAssignment,
     GroupMember,
+    OrganizationMembership,
     RoleScope,
     Scope,
     User,
@@ -376,11 +377,14 @@ class RBACService(BaseOrgService):
         # Verify group exists
         await self.get_group(group_id)
 
-        # Verify user exists in the org (check org membership table)
-        stmt = select(User).where(User.id == user_id)  # pyright: ignore[reportArgumentType]
+        # Verify user belongs to this organization
+        stmt = select(OrganizationMembership).where(
+            OrganizationMembership.user_id == user_id,
+            OrganizationMembership.organization_id == self.organization_id,
+        )
         result = await self.session.execute(stmt)
         if result.scalar_one_or_none() is None:
-            raise TracecatNotFoundError("User not found")
+            raise TracecatNotFoundError("User not found in organization")
 
         # Check if already a member
         stmt = select(GroupMember).where(
