+++ tracecat/authz/rbac/__init__.py
@@ -1,1 +1,7 @@
-"""RBAC (Role-Based Access Control) service and API module."""
+"""RBAC (Role-Based Access Control) OSS module.
+
+The user_scopes_router is provided here for the ScopeProvider in the frontend.
+
+Full RBAC management (roles, groups, scopes, assignments) is available in
+the Enterprise Edition (tracecat_ee.rbac).
+"""
